refactor: boot method
